Refactored Queries in React into Separate File
